pyproject.toml
[tool.poetry]
name = "qlora-sam"
version = "0.1.0"
description = "Fine-tune SAM with Huggigface PEFT and Trainer apply QLoRA, DoRA, etc."
authors = ["Magnus Samuelsen"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
transformers = "^4.44.2"
pycocotools = "^2.0.8"
numpy = "^2.1.1"
peft = "^0.12.0"
bitsandbytes = "^0.43.3"
accelerate = "^0.34.2"
evaluate = "^0.4.3"
tensorboardx = "^2.6.2.2"
tensorboard = "^2.17.1"
torchvision = "^0.19.1"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
